model/requirements.txt
opencv-python
mediapipe==0.10.9
numpy
fastapi
uvicorn
groq
python-dotenv
websockets
